Move source row to separate file
