+++ ui/pages/confirmations/components/transactions/intents-section/intents-section.tsx
@@ -52,11 +52,10 @@ import { NATIVE_TOKEN_ADDRESS } from '../../../../../helpers/constants/intents';
 import BigNumber from 'bignumber.js';
 import { useTokenFiatAmount } from '../../../../../hooks/useTokenFiatAmount';
 import { IntentsTargetRow } from '../intents-target-row/intents-target-row';
-
-type SelectedToken = {
-  address: Hex;
-  chainId: Hex;
-};
+import {
+  IntentsSourceRow,
+  SelectedToken,
+} from '../intents-source-row/intents-source-row';
 
 const log = createProjectLogger('intents');
 
@@ -197,178 +196,5 @@ function IntentsNetworkFeeRow({
     <ConfirmInfoRow label="Network Fee">
       <ConfirmInfoRowText text={`${networkFeeFiat} ${networkFee}`} />
     </ConfirmInfoRow>
-  );
-}
-
-function IntentsSourceRow({
-  loading,
-  onChange,
-  sourceTokenAmounts,
-  sourceTokenChainId,
-  sourceTokenAddress,
-  targetChainId,
-}: {
-  loading?: boolean;
-  onChange?: (token: SelectedToken) => void;
-  sourceTokenAmounts?: IntentSourceAmounts;
-  sourceTokenChainId: Hex;
-  sourceTokenAddress: Hex;
-  targetChainId: Hex;
-}) {
-  const [asset, setAsset] = useState<
-    AssetWithDisplayData<NativeAsset> | AssetWithDisplayData<ERC20Asset>
-  >({
-    chainId: targetChainId,
-    image: './images/eth_logo.svg',
-    type: AssetType.native,
-    symbol: 'ETH',
-  } as never);
-
-  const defaultNetwork = useSelector((state) =>
-    selectNetworkConfigurationByChainId(state, targetChainId),
-  );
-
-  const [network, setNetwork] = useState<NetworkConfiguration>(defaultNetwork);
-
-  useEffect(() => {
-    onChange?.({
-      address: (asset.address as Hex) || NATIVE_TOKEN_ADDRESS,
-      chainId: network.chainId,
-    });
-  }, [asset.address, network.chainId, onChange]);
-
-  const sourceAmountTotal = sourceTokenAmounts
-    ?.reduce(
-      (acc, amount) =>
-        acc.plus(new BigNumber(amount.sourceTokenAmountFormatted)),
-      new BigNumber(0),
-    )
-    .round(6)
-    .toString();
-
-  const sourceTokenFiat = useTokenFiatAmount(
-    sourceTokenAddress,
-    sourceAmountTotal,
-    undefined,
-    {},
-    true,
-    sourceTokenChainId,
-  );
-
-  return (
-    <ConfirmInfoRow label="Pay">
-      <Box
-        display={Display.Flex}
-        flexDirection={FlexDirection.Row}
-        justifyContent={JustifyContent.flexEnd}
-        alignItems={AlignItems.center}
-        gap={2}
-      >
-        {loading && (
-          <div role="progressbar">
-            <Preloader size={20} />
-          </div>
-        )}
-        <Text>{sourceTokenFiat}</Text>
-        <Text>{sourceAmountTotal}</Text>
-        <AssetPickerWrapper
-          asset={asset}
-          network={network}
-          onAssetChange={(newAsset) => setAsset(newAsset)}
-          onNetworkChange={(newNetwork) => setNetwork(newNetwork)}
-        />
-      </Box>
-    </ConfirmInfoRow>
-  );
-}
-
-function AssetPickerWrapper({
-  asset,
-  network,
-  onAssetChange,
-  onNetworkChange,
-}: {
-  asset: AssetWithDisplayData<NativeAsset> | AssetWithDisplayData<ERC20Asset>;
-  network: NetworkConfiguration;
-  onAssetChange: (
-    asset: AssetWithDisplayData<NativeAsset> | AssetWithDisplayData<ERC20Asset>,
-  ) => void;
-  onNetworkChange: (network: NetworkConfiguration) => void;
-}) {
-  const supportedChains = useSelector(getFromChains);
-
-  return (
-    <AssetPicker
-      asset={asset}
-      header="Select token"
-      isMultiselectEnabled={true}
-      visibleTabs={[TabName.TOKENS]}
-      onAssetChange={onAssetChange}
-      networkProps={{
-        network,
-        networks: supportedChains,
-        onNetworkChange: (network) =>
-          onNetworkChange(network as NetworkConfiguration),
-      }}
-    >
-      {(onClickHandler, networkImageSrc) => (
-        <AssetPickerOverride
-          asset={asset}
-          network={network}
-          networkImageSrc={networkImageSrc}
-          onClick={onClickHandler}
-        />
-      )}
-    </AssetPicker>
-  );
-}
-
-function AssetPickerOverride({
-  asset,
-  network,
-  networkImageSrc,
-  onClick,
-}: {
-  asset: AssetWithDisplayData<NativeAsset> | AssetWithDisplayData<ERC20Asset>;
-  network: NetworkConfiguration;
-  networkImageSrc?: string;
-  onClick?: () => void;
-}) {
-  return (
-    <Box
-      display={Display.Flex}
-      flexDirection={FlexDirection.Row}
-      borderRadius={BorderRadius.pill}
-      borderColor={BorderColor.borderMuted}
-      alignItems={AlignItems.center}
-      gap={2}
-      paddingInline={2}
-      onClick={onClick}
-      style={{
-        cursor: 'pointer',
-      }}
-    >
-      <BadgeWrapper
-        className="intents-network-icon-wrapper"
-        badge={
-          networkImageSrc ? (
-            <AvatarNetwork
-              className="intents-network-icon"
-              size={AvatarNetworkSize.Xs}
-              src={networkImageSrc}
-              name={network.name}
-            />
-          ) : undefined
-        }
-      >
-        <AvatarToken
-          borderRadius={BorderRadius.full}
-          src={asset.image}
-          size={AvatarTokenSize.Xs}
-        />
-      </BadgeWrapper>
-      <Text>{asset.symbol}</Text>
-      <Icon name={IconName.ArrowDown} size={IconSize.Sm} />
-    </Box>
   );
 }
